requirements.txt
rockman
Django==3.2.8

djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-api-key

django-environ==0.9.0
django-ckeditor==6.3.2

drf-extra-fields==3.7.0
drf-yasg[validation]==1.21.7

idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3

django-storages==1.12.3

mysqlclient

argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0

gunicorn==20.1.0
whitenoise==6.2.0

selenium~=4.7.2
pytz==2023.3
